pyproject.toml
[project]
name = "meshtastic-handler-server"
version = "0.1.0"
description = "A Python-based Meshtastic server with a plugin architecture for handling different types of requests"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Meshtastic Handler Team" }
]
keywords = ["meshtastic", "lora", "mesh", "radio", "plugin"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "meshtastic>=2.0",
    "httpx>=0.25",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "respx>=0.21",
]

[project.scripts]
meshtastic-handler = "meshtastic_handler.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meshtastic_handler"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
